requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Configuration
pydantic-settings==2.1.0

# Scheduler
apscheduler==3.10.4

# HTTP Client
httpx==0.26.0

# AI
openai==1.12.0

# PDF Generation
weasyprint==60.2
jinja2==3.1.3

# Email
resend==0.7.2

# Payments
yookassa==3.1.0

# RSS & HTML Parsing
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.1.0

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
